+++ FoodMap/src/main/java/com/example/FoodMap/componets/users/usersRepository.java
@@ -12,7 +12,7 @@ import org.springframework.stereotype.Repository;
 public interface usersRepository extends JpaRepository<users, Long> {
     
     @Query("select u from users u where u.id = ?1")
-    users getUser(int id);
+    users getUser(Long id);
 
     @Query("select u from users u where u.email = ?1")
     users findUsersByEmail(String email);
